feat: support for toggling between 12/24-hour time in UI
This is implemented as baseline for region and timezone selection.
