+++ web/src/hooks/useDayjs.ts
@@ -1,0 +1,6 @@
+import { DayjsContext } from '@/providers/DayjsProvider';
+import { useContext } from 'react';
+
+export const useDayjs = () => {
+  return useContext(DayjsContext).dayjs;
+};
